# perf: improved the performance of the critical function to escape text